+++ src/db/sql/tables/moves.sql
@@ -193,7 +193,7 @@ CREATE TABLE IF NOT EXISTS pmove_creates_field_state (
   pmove_generation_id TINYINT UNSIGNED NOT NULL,
   pmove_id SMALLINT UNSIGNED NOT NULL,
   field_state_generation_id TINYINT UNSIGNED NOT NULL,
-  field_state_id TINYINT UNSIGNED NOT NULL,
+  field_state_id SMALLINT UNSIGNED NOT NULL,
   turns TINYINT UNSIGNED NOT NULL,
 
   PRIMARY KEY (pmove_generation_id, pmove_id, field_state_generation_id, field_state_id),
@@ -211,7 +211,7 @@ CREATE TABLE IF NOT EXISTS pmove_removes_field_state (
   pmove_generation_id TINYINT UNSIGNED NOT NULL,
   pmove_id SMALLINT UNSIGNED NOT NULL,
   field_state_generation_id TINYINT UNSIGNED NOT NULL,
-  field_state_id TINYINT UNSIGNED NOT NULL,
+  field_state_id SMALLINT UNSIGNED NOT NULL,
 
   PRIMARY KEY (pmove_generation_id, pmove_id, field_state_generation_id, field_state_id),
   FOREIGN KEY (pmove_generation_id, pmove_id) REFERENCES pmove(generation_id, pmove_id)
